+++ src/main/java/com/twilio/rest/api/v2010/account/message/Feedback.java
@@ -16,6 +16,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.base.MoreObjects;
 import com.twilio.base.Resource;
 import com.twilio.converter.DateConverter;
+import com.twilio.converter.Promoter;
 import com.twilio.exception.ApiConnectionException;
 import com.twilio.exception.ApiException;
 import com.twilio.exception.RestException;
@@ -56,14 +57,7 @@ public class Feedback extends Resource {
          */
         @JsonCreator
         public static Outcome forValue(final String value) {
-            String normalized = value.replace("-", "_").toUpperCase();
-            try {
-                return Outcome.valueOf(normalized);
-            } catch (RuntimeException e) {
-
-                // Don't blow up of value does not exist
-                return null;
-            }
+            return Promoter.enumFromString(value, Outcome.values());
         }
     }