+++ src/main/java/com/twilio/rest/ipmessaging/v1/Credential.java
@@ -16,6 +16,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.base.MoreObjects;
 import com.twilio.base.Resource;
 import com.twilio.converter.DateConverter;
+import com.twilio.converter.Promoter;
 import com.twilio.exception.ApiConnectionException;
 import com.twilio.exception.ApiException;
 import com.twilio.exception.RestException;
@@ -58,14 +59,7 @@ public class Credential extends Resource {
          */
         @JsonCreator
         public static PushService forValue(final String value) {
-            String normalized = value.replace("-", "_").toUpperCase();
-            try {
-                return PushService.valueOf(normalized);
-            } catch (RuntimeException e) {
-
-                // Don't blow up of value does not exist
-                return null;
-            }
+            return Promoter.enumFromString(value, PushService.values());
         }
     }