+++ src/main/java/com/twilio/rest/messaging/v1/service/PhoneNumberDeleter.java
@@ -5,9 +5,9 @@
  *       /       /
  */
 
-package com.twilio.rest.api.v2010.account;
+package com.twilio.rest.messaging.v1.service;
 
-import com.twilio.base.Fetcher;
+import com.twilio.base.Deleter;
 import com.twilio.exception.ApiConnectionException;
 import com.twilio.exception.ApiException;
 import com.twilio.exception.RestException;
@@ -17,45 +17,41 @@ import com.twilio.http.Response;
 import com.twilio.http.TwilioRestClient;
 import com.twilio.rest.Domains;
 
-public class SandboxFetcher extends Fetcher<Sandbox> {
-    private String pathAccountSid;
+public class PhoneNumberDeleter extends Deleter<PhoneNumber> {
+    private final String pathServiceSid;
+    private final String pathSid;
 
     /**
-     * Construct a new SandboxFetcher.
-     */
-    public SandboxFetcher() {
-    }
-
-    /**
-     * Construct a new SandboxFetcher.
+     * Construct a new PhoneNumberDeleter.
      * 
-     * @param pathAccountSid The account_sid
+     * @param pathServiceSid The service_sid
+     * @param pathSid The sid
      */
-    public SandboxFetcher(final String pathAccountSid) {
-        this.pathAccountSid = pathAccountSid;
+    public PhoneNumberDeleter(final String pathServiceSid, 
+                              final String pathSid) {
+        this.pathServiceSid = pathServiceSid;
+        this.pathSid = pathSid;
     }
 
     /**
-     * Make the request to the Twilio API to perform the fetch.
+     * Make the request to the Twilio API to perform the delete.
      * 
      * @param client TwilioRestClient with which to make the request
-     * @return Fetched Sandbox
      */
     @Override
     @SuppressWarnings("checkstyle:linelength")
-    public Sandbox fetch(final TwilioRestClient client) {
-        this.pathAccountSid = this.pathAccountSid == null ? client.getAccountSid() : this.pathAccountSid;
+    public boolean delete(final TwilioRestClient client) {
         Request request = new Request(
-            HttpMethod.GET,
-            Domains.API.toString(),
-            "/2010-04-01/Accounts/" + this.pathAccountSid + "/Sandbox.json",
+            HttpMethod.DELETE,
+            Domains.MESSAGING.toString(),
+            "/v1/Services/" + this.pathServiceSid + "/PhoneNumbers/" + this.pathSid + "",
             client.getRegion()
         );
 
         Response response = client.request(request);
 
         if (response == null) {
-            throw new ApiConnectionException("Sandbox fetch failed: Unable to connect to server");
+            throw new ApiConnectionException("PhoneNumber delete failed: Unable to connect to server");
         } else if (!TwilioRestClient.SUCCESS.apply(response.getStatusCode())) {
             RestException restException = RestException.fromJson(response.getStream(), client.getObjectMapper());
             if (restException == null) {
@@ -71,6 +67,6 @@ public class SandboxFetcher extends Fetcher<Sandbox> {
             );
         }
 
-        return Sandbox.fromJson(response.getStream(), client.getObjectMapper());
+        return response.getStatusCode() == 204;
     }
 }