+++ src/main/java/com/twilio/rest/notify/v1/service/Notification.java
@@ -17,6 +17,7 @@ import com.google.common.base.MoreObjects;
 import com.twilio.base.Resource;
 import com.twilio.converter.Converter;
 import com.twilio.converter.DateConverter;
+import com.twilio.converter.Promoter;
 import com.twilio.exception.ApiConnectionException;
 import com.twilio.exception.ApiException;
 import com.twilio.exception.RestException;
@@ -58,14 +59,7 @@ public class Notification extends Resource {
          */
         @JsonCreator
         public static Priority forValue(final String value) {
-            String normalized = value.replace("-", "_").toUpperCase();
-            try {
-                return Priority.valueOf(normalized);
-            } catch (RuntimeException e) {
-
-                // Don't blow up of value does not exist
-                return null;
-            }
+            return Promoter.enumFromString(value, Priority.values());
         }
     }