+++ src/main/java/com/twilio/rest/preview/marketplace/InstalledAddOn.java
@@ -30,6 +30,7 @@ import org.joda.time.DateTime;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.URI;
+import java.util.List;
 import java.util.Map;
 import java.util.Objects;