+++ src/main/java/com/twilio/rest/preview/sync/service/DocumentCreator.java
@@ -18,6 +18,7 @@ import com.twilio.http.Response;
 import com.twilio.http.TwilioRestClient;
 import com.twilio.rest.Domains;
 
+import java.util.List;
 import java.util.Map;
 
 public class DocumentCreator extends Creator<Document> {